% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly_functions.R
\name{expand_matrix_text}
\alias{expand_matrix_text}
\title{Create a term-by-term matrix.}
\usage{
expand_matrix_text(text_1, text_2, stop_words = NULL,
  make_lower = TRUE, rem_punct = TRUE, rem_num = FALSE)
}
\arguments{
\item{text_1}{The first list of words or document input.}

\item{text_2}{The second list of words or document input.}

\item{stop_words}{A list of stop words to be used. Defaults to no stop words used (NULL).}

\item{make_lower}{Should text be converted to lower case? Defaults to TRUE.}

\item{rem_punct}{Should punctuation be removed? Defaults to TRUE.}

\item{rem_num}{Should numerics be removed? Defaults to TRUE.}
}
\description{
Takes in two documents or two lists of words and creates a matrix of all possible word-by-word combinations. Used primarily by 'leven_dist' function.
}
\examples{
expand_matrix_text(c("duck", "pig", "cat"), c("dog", "cat"))
expand_matrix_text("HEllo I'm me", "hI tHeRe")
}
\keyword{leven_dist}
