% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly_functions.R
\name{leven_dist}
\alias{leven_dist}
\title{Levenshtein distance.}
\usage{
leven_dist(mat_in, col_num_1 = 1, col_num_2 = 2)
}
\arguments{
\item{mat_in}{The output from the 'expand_matrix_text' function.}

\item{col_num_1}{The numerical column index for the first column of words.}

\item{col_num_2}{The numerical column index for the second column of words.}
}
\description{
Takes in output from the 'expand_matrix_text' function and returns the Levenshtein distance between each word-by-word combination.
}
\examples{
leven_dist(expand_matrix_text(c("duck", "pig", "cat"), c("dog", "cat")))
leven_dist(expand_matrix_text("HEllo I'm me", "hI tHeRe"))
}
\keyword{expand_matrix_text}
