% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly_functions.R
\name{make_iterator}
\alias{make_iterator}
\title{Returns an iterator. Used by prepare_dtm() function.}
\usage{
make_iterator(dat_in, text_var_name, id_var_name, see_verbose = TRUE)
}
\arguments{
\item{dat_in}{The data.frame or tibble input.}

\item{text_var_name}{The name of the column in dat_in containing the text source.}

\item{id_var_name}{The name of the column in dat_in containing the id source.}

\item{see_verbose}{Akin to verbose. Defaults to TRUE.}
}
\description{
Creates iterator object to be passed to make_vectorizer(). Ultimately used by prepare_dtm() to construct a DTM.
}
\examples{
iter_ret <- product_review_df \%>\% make_iterator("product_review", "product_id")
}
\keyword{make_vectorizer}
\keyword{prepare_dtm,}
