% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly_functions.R
\name{make_vectorizer}
\alias{make_vectorizer}
\title{Returns a vectorizer. Used by prepare_dtm() function.}
\usage{
make_vectorizer(iterator_in, stop_word_list, n_gram_min, n_gram_max,
  see_verbose = TRUE)
}
\arguments{
\item{iterator_in}{The iterator output from make_iterator().}

\item{stop_word_list}{A list of stopwords.}

\item{n_gram_min}{Minimum n-gram for vocabulary.}

\item{n_gram_max}{Maximum n-gram for vocabulary.}

\item{see_verbose}{Akin to verbose. Defaults to TRUE.}
}
\description{
Takes in iterator from make_iterator() and returns a vectorizer. Ultimately used by prepare_dtm() to construct a DTM.
}
\examples{
vectorizer_ret <- iter_ret \%>\% make_vectorizer()
vectorizer_ret <- iter_ret \%>\% make_vectorizer(stop_word_list = my_stopword_list)
}
\keyword{make_iterator}
\keyword{prepare_dtm,}
