% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly_functions.R
\name{prepare_dtm}
\alias{prepare_dtm}
\title{Create a document-term-matrix (DTM) for NLP analysis and modeling.}
\usage{
prepare_dtm(dat_in, text_col_name, return_vectorizer = FALSE,
  use_vectorizer = NULL, stopword_list = c("the", "and", "etc"),
  vect_n_gram_min = 1, vect_n_gram_max = 1, see_verbose = TRUE)
}
\arguments{
\item{dat_in}{The data.frame or tibble input data.}

\item{text_col_name}{The name of the column in dat_in containing the text source.}

\item{return_vectorizer}{Should the created vectorizer be returned? Only set this to TRUE if input data is TRAIN data, otherwise keep FALSE (default).}

\item{use_vectorizer}{The vectorizer to be used. Only provide arg if input data is TEST data, otherwise keep NULL (default).}

\item{stopword_list}{A list of stopwords.}

\item{vect_n_gram_min}{Minimum n-gram for vocabulary.}

\item{vect_n_gram_max}{Maximum n-gram for vocabulary.}

\item{see_verbose}{Akin to verbose. Defaults to TRUE.}
}
\description{
Allows for creation of a document-term-matrix (DTM), returned in sparse matrix format.
}
\examples{
dtm_train_vect <- train_ \%>\% prepare_dtm("product_review", return_vectorizer = TRUE)
dtm_train <- dtm_train_vect$"dtm_obj"
dtm_test <- test_ \%>\% prepare_dtm("product_review", use_vectorizer = dtm_train_vect$"vectorizer_out")
}
\keyword{make_iterator,}
\keyword{make_vectorizer}
