% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly_functions.R
\name{show_diffr}
\alias{show_diffr}
\title{Show the difference between two vectors of words.}
\usage{
show_diffr(word_vec_1, word_vec_2, title_1 = "Text 1",
  title_2 = "Text 2")
}
\arguments{
\item{word_vec_1}{The first vector of words to be compared.}

\item{word_vec_2}{The second vector of words to be compared.}

\item{title_1}{The title given to the first vector of words. Defaults to "Text 1".}

\item{title_2}{The title given to the second vector of words. Defaults to "Text 2".}
}
\description{
Given two vectors of words, this function uses the diffr lib and displays a comparison between them.
}
\examples{
show_diffr(c("cat", "dog", "puppy"), c("cat", "dog", "Puppy", "kitten"))
}
\keyword{diffr}
