% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly_functions.R
\name{tryel}
\alias{tryel}
\title{A more intuitive take on R's try function.}
\usage{
tryel(expr, el, return_error = FALSE)
}
\arguments{
\item{expr}{The expression to be attempted.}

\item{el}{The return if expr fails. Can be a value or alternate function.}

\item{return_error}{Should the error message of a failed expr be returned? Defaults to FALSE.}
}
\description{
Combines the benefits of try and else into one easy-to-use function.
}
\examples{
bad_fun <- function() log("potato")
tryel(bad_fun(), 100)
good_fun <- function() log(10)
tryel(bad_fun(), good_fun())
}
\keyword{else,}
\keyword{ifelse}
\keyword{try,}
