% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordly-package.R
\docType{package}
\name{wordly-package}
\alias{wordly}
\alias{wordly-package}
\title{wordly: Natural Language Processing (NLP) functionality and modeling}
\description{
This package provides functions for NLP analysis and modeling. It is mainly a driver for existing libs such as text2vec, tokenizers, etc.
}
\author{
\strong{Maintainer}: Tom A \email{tom@fake-email-do-not-use.com} (YOUR-ORCID-ID)

}
\keyword{internal}
